requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.20
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
